refactor(calibration-storage): rename calibration files to more sensible names
'get', 'delete' etc can be rather confusing when using the calibration_storage module. Instead, we
should name each file based on the calibration data they handle.
