+++ api/src/opentrons/calibration_storage/ot3/__init__.py
@@ -1,3 +1,4 @@
-from . import *
+from . import deck_attitude as ot3_deck_attitude, pipette_offset as ot3_pipette_offset, tip_length as ot3_tip_length, schemas as ot3_schemas, gripper_offset as ot3_gripper_offset
 
-__all__ = ["ot3_calibration_storage"]
+
+__all__ = ["ot3_deck_attitude", "ot3_pipette_offset", "ot3_tip_length", "ot3_schemas", "ot3_gripper_offset"]
