+++ api/src/opentrons/hardware_control/instruments/ot3/instrument_calibration.py
@@ -7,10 +7,9 @@ from opentrons.config.robot_configs import (
     default_pipette_offset,
     default_gripper_calibration_offset,
 )
-from opentrons.types import Mount, Point
-from opentrons.calibration_storage import types as cal_top_types
+from opentrons.types import Point
+from opentrons.calibration_storage import types as cal_top_types, ot3_pipette_offset, ot3_gripper_offset
 from opentrons.hardware_control.types import OT3Mount
-from opentrons.calibration_storage.ot3 import get
 
 
 @dataclass
@@ -56,7 +55,7 @@ def load_pipette_offset(
     else:
         checked_mount = mount
     if pip_id:
-        pip_offset_data = get.get_pipette_offset(
+        pip_offset_data = ot3_pipette_offset.get_pipette_offset(
             typing.cast(cal_top_types.PipetteId, pip_id), checked_mount
         )
         if pip_offset_data:
@@ -83,7 +82,7 @@ def load_gripper_calibration_offset(
         status=cal_top_types.CalibrationStatus(),
     )
     if gripper_id and ff.enable_ot3_hardware_controller():
-        grip_offset_data = get.get_gripper_calibration_offset(
+        grip_offset_data = ot3_gripper_offset.get_gripper_calibration_offset(
             typing.cast(cal_top_types.GripperId, gripper_id)
         )
         if grip_offset_data:
