+++ api/tests/opentrons/hardware_control/test_calibration_functions.py
@@ -1,9 +1,8 @@
 import numpy as np
 
 from opentrons import config
-from opentrons.calibration_storage import file_operators as io
+from opentrons.calibration_storage import file_operators as io, ot2_pipette_offset
 
-from opentrons.calibration_storage.ot2 import modify, cache as calibration_cache
 from opentrons.hardware_control import robot_calibration
 from opentrons.hardware_control.instruments.ot2 import instrument_calibration
 from opentrons.util.helpers import utc_now
@@ -57,7 +56,6 @@ def test_load_calibration(ot_config_tempdir):
         "last_modified": utc_now(),
         "tiprack": "hash",
     }
-    calibration_cache._deck_calibration.cache_clear()
     io.save_to_file(pathway, data)
     obj = robot_calibration.load_attitude_matrix()
     transform = [[1, 0, 1], [0, 1, -0.5], [0, 0, 1]]
@@ -74,7 +72,6 @@ def test_load_malformed_calibration(ot_config_tempdir):
         "tiprack": "hash",
         "statu": [1, 2, 3],
     }
-    calibration_cache._deck_calibration.cache_clear()
     io.save_to_file(pathway, data)
     obj = robot_calibration.load_attitude_matrix()
     assert np.allclose(obj.attitude, [[1, 0, 0], [0, 1, 0], [0, 0, 1]])
@@ -92,7 +89,7 @@ def test_load_pipette_offset(ot_config_tempdir):
     mount = Mount.LEFT
     offset = Point(1, 2, 3)
 
-    modify.save_pipette_calibration(
+    ot2_pipette_offset.save_pipette_calibration(
         offset, pip_id, mount, "hash", "opentrons/opentrons_96_tiprack_10ul/1"
     )
     obj = instrument_calibration.load_pipette_offset(pip_id, mount)
@@ -101,7 +98,6 @@ def test_load_pipette_offset(ot_config_tempdir):
 
 
 def test_load_bad_pipette_offset(ot_config_tempdir):
-    calibration_cache._pipette_offset_calibrations.cache_clear()
     path = config.get_opentrons_path("pipette_calibration_dir") / "left"
     path.mkdir(parents=True, exist_ok=True)
     calpath = path / "fakePip.json"
