+++ robot-server/robot_server/robot/calibration/util.py
@@ -8,8 +8,7 @@ from opentrons.hardware_control.types import CriticalPoint
 from opentrons.protocol_api import labware
 from opentrons.protocols.geometry import planning
 from opentrons.protocols.geometry.deck import Deck
-from opentrons.calibration_storage import helpers, types as cal_types
-from opentrons.calibration_storage.ot2 import modify
+from opentrons.calibration_storage import helpers, types as cal_types, ot2_tip_length
 from opentrons.types import Point, Location
 
 from robot_server.service.errors import RobotServerError
@@ -200,10 +199,10 @@ def save_tip_length_calibration(
     # TODO: 07-22-2020 parent slot is not important when tracking
     # tip length data, hence the empty string, we should remove it
     # from create_tip_length_data in a refactor
-    tip_length_data = modify.create_tip_length_data(
+    tip_length_data = ot2_tip_length.create_tip_length_data(
         tip_rack._implementation.get_definition(), tip_length_offset
     )
-    modify.save_tip_length_calibration(
+    ot2_tip_length.save_tip_length_calibration(
         cast(cal_types.PipetteId, pipette_id), tip_length_data
     )
