+++ robot-server/robot_server/service/pipette_offset/router.py
@@ -3,8 +3,7 @@ from fastapi import APIRouter
 from typing import Optional, cast
 
 from opentrons import types as ot_types
-from opentrons.calibration_storage import types as cal_types
-from opentrons.calibration_storage.ot2 import get as get_cal, delete, schemas
+from opentrons.calibration_storage import types as cal_types, ot2_pipette_offset, ot2_schemas
 
 from robot_server.errors import ErrorBody
 from robot_server.service.pipette_offset import models as pip_models
@@ -15,7 +14,7 @@ router = APIRouter()
 
 
 def _format_calibration(
-    calibration: schemas.v1.PipetteOffsetCalibration,
+    calibration: ot2_schemas.v1.PipetteOffsetCalibration,
 ) -> pip_models.PipetteOffsetCalibration:
     # TODO (lc 09-20-2022) We should use the calibration
     # status model in calibration storage.
@@ -49,7 +48,7 @@ async def get_all_pipette_offset_calibrations(
     pipette_id: Optional[str] = None, mount: Optional[pip_models.MountType] = None
 ) -> pip_models.MultipleCalibrationsResponse:
 
-    all_calibrations = get_cal.get_all_pipette_offset_calibrations()
+    all_calibrations = ot2_pipette_offset.get_all_pipette_offset_calibrations()
     if not all_calibrations:
         return pip_models.MultipleCalibrationsResponse(
             data=[_format_calibration(cal) for cal in all_calibrations],
@@ -79,7 +78,7 @@ async def delete_specific_pipette_offset_calibration(
     pipette_id: str, mount: pip_models.MountType
 ):
     try:
-        delete.delete_pipette_offset_file(
+        ot2_pipette_offset.delete_pipette_offset_file(
             cast(cal_types.PipetteId, pipette_id), ot_types.Mount[mount.upper()]
         )
     except FileNotFoundError:
