+++ robot-server/robot_server/service/tip_length/router.py
@@ -2,8 +2,7 @@ from starlette import status
 from fastapi import APIRouter
 from typing import Optional, cast
 
-from opentrons.calibration_storage import types as cal_types
-from opentrons.calibration_storage.ot2 import get as get_cal, delete, schemas
+from opentrons.calibration_storage import types as cal_types, ot2_tip_length, ot2_schemas
 
 from robot_server.errors import ErrorBody
 from robot_server.service.tip_length import models as tl_models
@@ -15,7 +14,7 @@ router = APIRouter()
 
 
 def _format_calibration(
-    calibration: schemas.v1.TipLengthCalibration,
+    calibration: ot2_schemas.v1.TipLengthCalibration,
 ) -> tl_models.TipLengthCalibration:
     # TODO (lc 09-20-2022) We should use the calibration
     # status model in calibration storage.
@@ -49,7 +48,7 @@ async def get_all_tip_length_calibrations(
     pipette_id: Optional[str] = None,
     tiprack_uri: Optional[str] = None,
 ) -> tl_models.MultipleCalibrationsResponse:
-    all_calibrations = get_cal.get_all_tip_length_calibrations()
+    all_calibrations = ot2_tip_length.get_all_tip_length_calibrations()
     print(all_calibrations)
     if not all_calibrations:
         return tl_models.MultipleCalibrationsResponse(
@@ -82,7 +81,7 @@ async def get_all_tip_length_calibrations(
 )
 async def delete_specific_tip_length_calibration(tiprack_hash: str, pipette_id: str):
     try:
-        delete.delete_tip_length_calibration(
+        ot2_tip_length.delete_tip_length_calibration(
             cast(cal_types.TiprackHash, tiprack_hash),
             cast(cal_types.PipetteId, pipette_id),
         )
