+++ robot-server/tests/robot/calibration/deck/test_user_flow.py
@@ -2,8 +2,7 @@ import pytest
 import datetime
 from mock import MagicMock, call
 from typing import List, Tuple
-from opentrons.calibration_storage.ot2 import schemas
-from opentrons.calibration_storage import types as cal_types
+from opentrons.calibration_storage import types as cal_types, ot2_schemas
 from opentrons.types import Mount, Point
 from opentrons.hardware_control.instruments.ot2 import pipette
 from opentrons.config import robot_configs
@@ -22,7 +21,7 @@ from robot_server.robot.calibration.deck.constants import (
 )
 
 
-PIP_OFFSET = schemas.v1.InstrumentOffsetSchema(
+PIP_OFFSET = ot2_schemas.v1.InstrumentOffsetSchema(
     offset=robot_configs.defaults_ot2.DEFAULT_PIPETTE_OFFSET,
     tiprack="some_tiprack",
     uri="custom/some_tiprack/1",
