+++ robot-server/tests/robot/calibration/tip_length/test_user_flow.py
@@ -8,8 +8,7 @@ from opentrons.hardware_control.instruments.ot2 import pipette
 from opentrons.protocol_api.labware import get_labware_definition
 from opentrons.config.pipette_config import load
 from opentrons.util.helpers import utc_now
-from opentrons.calibration_storage.ot2 import schemas
-from opentrons.calibration_storage import types as cal_types
+from opentrons.calibration_storage import types as cal_types, ot2_schemas
 
 from robot_server.service.errors import RobotServerError
 from robot_server.service.session.models.command_definitions import CalibrationCommand
@@ -17,7 +16,7 @@ from robot_server.robot.calibration.tip_length.user_flow import TipCalibrationUs
 
 stub_jog_data = {"vector": Point(1, 1, 1)}
 
-PIP_CAL = schemas.v1.InstrumentOffsetSchema(
+PIP_CAL = ot2_schemas.v1.InstrumentOffsetSchema(
     offset=[0, 0, 0],
     tiprack="some_tiprack",
     uri="custom/some_tiprack/1",
